requirements.txt
PySide6
qt-material
configparser
grpcio
grpcio-tools
adafruit-circuitpython-dht